% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msoc.R
\docType{package}
\name{msoc-package}
\alias{msoc-package}
\title{msoc}
\description{
Package to make the \code{msoc} library (\verb{https://github.com/herumi/msoffice})
for encrypting/decrypting of Office Open XML files available under R.
}
\details{
The password is expected to be plain text. If security is of importance,
do not use this package. If the password is lost, opening the file will
be impossible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://janmarvin.github.io/msoc/}
  \item \url{https://github.com/JanMarvin/msoc}
  \item Report bugs at \url{https://github.com/JanMarvin/msoc/issues}
}

}
\author{
\strong{Maintainer}: Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}

Authors:
\itemize{
  \item Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}
}

Other contributors:
\itemize{
  \item Shigeo Mitsunari (Author of included msoc library) [copyright holder]
}

}
