% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msoc.R
\name{msoc}
\alias{msoc}
\alias{encrypt}
\alias{decrypt}
\title{Encryption/Decryption function}
\usage{
encrypt(input, output = NULL, pass)

decrypt(input, output = NULL, pass)
}
\arguments{
\item{input}{input file}

\item{output}{(optional) output file. If none is provided, construct a
temporary output file with the file extension of the input file}

\item{pass}{a password to decrypt/encrypt the input file. The password is
expected to be plain text. If security is of importance, do not use this
package. If the password is lost, opening the file will be impossible.}
}
\value{
a path to the output file. Either specified or temporary
}
\description{
Encryption/Decryption function
}
